# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Schedurx Backend API
  version: 1.0.0
  description: >
    Clinic business API — patients, doctors, appointments, and calendar tools.
    The /tools/* paths are called directly by the Ultravox voice agent during a
    live call; /internal/* is called only by schedurx-ultravox-demo-api.
servers:
  - url: https://backend.example.test
    description: Replace with this service's public URL for deployed environments.
components:
  securitySchemes:
    ToolsBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: TOOLS_API_KEY — configured into Ultravox's tool definitions by the demo bridge.
    InternalBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: INTERNAL_API_KEY — held only by schedurx-ultravox-demo-api.
    FirebaseBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A Firebase ID token for a Staff account. Custom claims (role, clinicId,
        doctorId) are set at onboarding time via POST /internal/staff and read
        by every /api/v1/* route — clinicId is always sourced from these claims,
        never from the request itself.
    NettuWebhookKeyAuth:
      type: apiKey
      in: header
      name: nettu-scheduler-webhook-key
      description: >
        Static key set once via `PUT /api/v1/account/webhook` on the nettu
        account (see scripts/setup-nettu-webhook.js) — nettu echoes this exact
        value back on every reminder callback. Not a bearer token.
    TwilioSignatureAuth:
      type: apiKey
      in: header
      name: X-Twilio-Signature
      description: >
        HMAC-SHA1 signature over the exact request URL + form params, verified
        against TWILIO_AUTH_TOKEN (see twilio-client.js's validateSignature).
        Twilio computes and sends this automatically; it isn't a credential a
        caller sets manually.
  schemas:
    SimpleError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Doctor:
      type: object
      required: [doctorId, fullName, specialty, qualification, languages, feeInr]
      properties:
        doctorId:
          type: string
        fullName:
          type: string
        specialty:
          type:
            - string
            - "null"
        qualification:
          type:
            - string
            - "null"
        languages:
          type:
            - array
            - string
            - "null"
          items:
            type: string
        feeInr:
          type:
            - number
            - integer
            - "null"
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SimpleError"
          examples:
            unauthorized:
              value:
                error: Unauthorized
    InternalServerError:
      description: Unhandled server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SimpleError"
          examples:
            internal:
              value:
                error: Internal server error
paths:
...
  /tools/doctors/list:
    post:
      summary: List active doctors for the resolved clinic.
      security:
        - ToolsBearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ultravoxCallId:
                  type: string
      responses:
        "200":
          description: Doctor list.
          content:
            application/json:
              schema:
                type: object
                required: [doctors]
                properties:
                  doctors:
                    type: array
                    items:
                      $ref: "#/components/schemas/Doctor"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Clinic missing from call context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /tools/doctors/select:
    post:
      summary: Resolve a spoken doctor name and store it in call context.
      security:
        - ToolsBearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                ultravoxCallId:
                  type: string
                doctorName:
                  type: string
      responses:
        "200":
          description: Doctor selected.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: No doctor matched.
          content:
            application/json:
              schema:
                type: object
                required: [error, availableDoctors]
                properties:
                  error:
                    type: string
                  availableDoctors:
                    type: array
                    items:
                      type: string
        "422":
          description: Missing doctorName or clinic context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimpleError"
        "500":
          $ref: "#/components/responses/InternalServerError"
...